backend/requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
pyjwt==2.8.0
werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.1 